creat backup

+++ FINAL/travel-booking/src/main/java/com/example/travelbooking/repository/FacilityRepository.java
@@ -3,5 +3,5 @@ package com.example.travelbooking.repository;
 import com.example.travelbooking.entity.Facility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacilityRepository extends JpaRepository<Facility, Long> {
+public interface FacilityRepository extends JpaRepository<Facility, Integer> {
 }
